requirements.txt
asn1crypto==1.5.1
bcrypt==4.3.0
click==8.2.1
dotenv==0.9.9
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
greenlet==3.2.3
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pip==25.1.1
psycopg2-binary==2.9.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
scramp==1.4.6
six==1.17.0
sqlparse==0.5.3
Werkzeug==3.1.3